+++ src/Dto/Playlist.php
@@ -6,7 +6,7 @@ namespace App\Dto;
 
 use App\Dto\Trait\BlameableTrait;
 use App\Dto\Trait\IdentifiableTrait;
-use App\Dto\Trait\RelationsModifiedTrait;
+use App\Dto\Trait\RelationsChecksumTrait;
 use App\Dto\Trait\TimestampableTrait;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
@@ -17,7 +17,7 @@ class Playlist
     use BlameableTrait;
     use IdentifiableTrait;
     use TimestampableTrait;
-    use RelationsModifiedTrait;
+    use RelationsChecksumTrait;
 
     #[Groups(['playlist-screen-region:read', 'screen-campaigns:read', 'campaigns/screens:read', 'slides/playlists:read', 'screen-groups/campaigns:read', 'campaigns/screen-groups:read'])]
     public string $title = '';