+++ src/Dto/Screen.php
@@ -6,7 +6,7 @@ namespace App\Dto;
 
 use App\Dto\Trait\BlameableTrait;
 use App\Dto\Trait\IdentifiableTrait;
-use App\Dto\Trait\RelationsModifiedTrait;
+use App\Dto\Trait\RelationsChecksumTrait;
 use App\Dto\Trait\TimestampableTrait;
 use Symfony\Component\Serializer\Annotation\Groups;
 
@@ -15,7 +15,7 @@ class Screen
     use BlameableTrait;
     use IdentifiableTrait;
     use TimestampableTrait;
-    use RelationsModifiedTrait;
+    use RelationsChecksumTrait;
 
     #[Groups(['campaigns/screens:read', 'screen-groups/screens:read'])]
     public string $title = '';